// src/pages/Diet/types/Diet.types.ts
export interface DietType {
    CALCIUM_QY: number // 칼슘
    CALORIE_QY: number // 칼로리
    CARBOH_QY: number // 탄수화물
    CELLU_QY: number // 섬유소
    COOK_MTH_CONT: string // 조리법
    FAT_QY: number // 지방
    FE_QY: number
    MATRL_NM: string // 재료
    MEAL_CLSF_NM: string
    MEAL_NM: string // 식단이름
    MEAL_PICTR_FILE_NM: string // 식단사진파일명
    NATRIUM_QY: number // 나트륨
    NIACIN_QY: number // 나이아신
    PHOSPH_QY: number // 인
    POTASSIUM_QY: number // 포타슘
    PROTEIN_QY: number // 단백질
    RIBOFLAMIN_QY: number  // 칼륨
    THIAMIN_QY: number  // 티아민
    VITAMINA_QY: number // 비타민A
    VITAMINC_QY: number // 비타민C
    id: number
}